flow-monitor: (fixes #2545) Optimized build fails for flow-monitor (2)

+++ src/flow-monitor/model/flow-probe.h
@@ -100,7 +100,7 @@ public:
   /// \param os the output stream
   /// \param indent number of spaces to use as base indentation level
   /// \param index FlowProbe index
-  void SerializeToXmlStream (std::ostream &os, int indent, uint32_t index) const;
+  void SerializeToXmlStream (std::ostream &os, uint16_t indent, uint32_t index) const;
 
 protected:
   Ptr<FlowMonitor> m_flowMonitor; //!< the FlowMonitor instance
