+++ src/flow-monitor/model/histogram.h
@@ -109,7 +109,7 @@ public:
    * \param indent number of spaces to use as base indentation level
    * \param elementName name of the element to serialize.
    */
-  void SerializeToXmlStream (std::ostream &os, int indent, std::string elementName) const;
+  void SerializeToXmlStream (std::ostream &os, uint16_t indent, std::string elementName) const;
 
 
 private:
